pyproject.toml
[project]
name = "taiwan-geodoc-hub"
version = "0.1.0"
description = ""

requires-python = ">=3.13,<3.14"
dependencies = [
  "injector",
  "pymupdf",
  "python-dotenv",
  "aiofiles",
  "httpx",
  "reactivex",
  "prefect",
  "google-cloud-pubsub",
  "google-cloud-run",
  "pybase62",
  "marshmallow",
]

[project.optional-dependencies]
firebase = [
  "starlette",
  "vellox",
  "firebase_functions",
  "asgi-lifespan",
]

cli = [
  "firebase-admin",
  "typer",
  "playwright",
  "uvicorn",
  "starlette",
]

dev = [
  "pytest",
  "pytest-testdox",
  "pytest-dotenv",
  "pytest-asyncio",
]

[tool.pytest.ini_options]
addopts = [
  "-v",
  "-s",
  "--testdox",
  "--capture=no",
  "--tb=long",
]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::pytest.PytestAssertRewriteWarning",
  "ignore:builtin type SwigPyPacked has no __module__ attribute",
  "ignore:builtin type SwigPyObject has no __module__ attribute",
  "ignore:builtin type swigvarlink has no __module__ attribute",
  "ignore::DeprecationWarning:reactivex.*",
]
norecursedirs = [
  ".*",
  "build",
  "dist",
  "*.egg",
  "info",
]
python_files = [
  "test_*.py",
]
pythonpath = ["./src"]
